+++ src/cli/utils/discover.ts
@@ -4,29 +4,10 @@
 
 import path from "path";
 
-// Support both ESM and CommonJS glob exports across versions
-import * as glob from "glob";
-const globSync =
-    (glob as unknown as { globSync?: typeof import("glob").globSync })
-        .globSync ??
-    (glob as unknown as { sync?: typeof import("glob").globSync }).sync ??
-    (
-        glob as unknown as {
-            default?: { globSync?: typeof import("glob").globSync };
-        }
-    ).default?.globSync ??
-    (glob as unknown as { default?: { sync?: typeof import("glob").globSync } })
-        .default?.sync;
-
-if (!globSync) {
-    throw new Error(
-        "Unable to resolve globSync from 'glob'. Please ensure a compatible glob version is installed.",
-    );
-}
-
 import storyblokConfig, { SCHEMA } from "../../config/config.js";
 import { buildOnTheFly } from "../../rollup/build-on-the-fly.js";
 import { getFileContentWithRequire, readFile } from "../../utils/files.js";
+import { safeGlobSync as globSync } from "../../utils/glob-utils.js";
 import {
     normalizeDiscover,
     filesPattern,